sistema completo
