+++ src/app/components/listarproductos/listarproductos.component.html
@@ -1,4 +1,4 @@
-<app-navbar></app-navbar>
+<app-navbar (ESubcategoria)="onRecibirSub($event)"></app-navbar>
 <div class="flex justify-between ">
   <div class="w-full">
     <header class="flex items-center  justify-between w-full px-4 pt-2 m-4 rounded-xl bg-yellow-400">
