+++ src/app/components/navbar/navbar.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input } from '@angular/core';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { combineLatest } from 'rxjs';
 import { Subcategorias } from 'src/app/interfaces/subcategorias';
@@ -12,8 +12,9 @@ import { SubcategoriasService } from 'src/app/services/subcategorias.service';
   styleUrls: ['./navbar.component.scss']
 })
 export class NavbarComponent {
-  @Input() categoria?: string;
-  @Input() subcategorias: Subcategorias[] = [];
+  @Output() ESubcategoria = new EventEmitter<string>();
+  categoria?: string;
+  subcategorias: Subcategorias[] = [];
   admin?: boolean;
   otro: any;
 
@@ -23,23 +24,36 @@ export class NavbarComponent {
     private categoriasvc: CategoriasService,
     private rutaactiva: ActivatedRoute
   ) {
-    if (this.categoria) {
-      console.log("existe categoria");
-
-    } else {
-      console.log("no hay categoria");
-
-    }
   }
 
   ngOnInit() {
     this.rutaactiva.params.subscribe(parametro => {
-      console.log("parametro", parametro);
-      
+      combineLatest(
+        [this.subcategoriassvc.getSubcategorias(),
+        this.iconosvc.getIconos(), this.categoriasvc.getCategorias()]
+      ).subscribe(([subcategorias, iconos, categorias]) => {
+        subcategorias.filter(x => x.categoria == parametro['categoria'])
+        let filtrado: Subcategorias[];
+        if (parametro['categoria']) {
+          filtrado = subcategorias.filter(x => x.categoria == parametro['categoria'])
+        } else {
+          filtrado = subcategorias;
+        }
+        this.subcategorias = filtrado.map(m => {
+          let icon = iconos.find(x => x.id == m.icono);
+          let cat = categorias.find(x => x.id == m.categoria);
+
+          return {
+            id: m.id,
+            nombre: m.nombre,
+            icono: icon ? icon['img'] : '',
+            categoria: cat ? cat['nombre'] : '',
+          }
+        })
+      })
+
     })
-
   }
-
 
   async onClickDelete(registro: Subcategorias) {
     const response = await this.subcategoriassvc.deleteSubcategorias(registro);
@@ -50,6 +64,10 @@ export class NavbarComponent {
     const response = await this.subcategoriassvc.updateSubcategorias(registro);
     console.log(response);
 
+  }
+
+  onEnviarSubCategoria(id: string ) {
+    this.ESubcategoria.emit(id)
   }
 
 }
